Rework PrintingWriter to include wrapper
